+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feat(notifications): add `progress` and `progress_weight` properties to `Notification` object and show the progress on the header of the app
 - feat(core): show the progress of the update using the new `progress` property of the `Notification` object
 - fix(camera): render the viewfinder on the display even if the display is paused - closes #78
+- refactor(core): make the power-off menu, a sub-menu with power-off and reboot action items - closes #123
 
 ## Version 0.15.4